remove player from suspend tracker when they leave
