+++ QSB/SuspendableSync/SuspendChangeEvent.cs
@@ -1,8 +1,4 @@
-﻿using System.Linq;
-using QSB.Events;
-using QSB.Player;
-using QSB.Utility;
-using QuantumUNET;
+﻿using QSB.Events;
 using QuantumUNET.Components;
 
 namespace QSB.SuspendableSync
@@ -35,50 +31,7 @@ namespace QSB.SuspendableSync
 				return;
 			}
 
-			var unsuspendedFor = SuspendableManager._unsuspendedFor[message.Identity];
-
-			var suspended = !unsuspendedFor.Contains(message.FromId);
-			if (message.Suspended == suspended)
-			{
-				return;
-			}
-
-			if (!message.Suspended)
-			{
-				unsuspendedFor.Add(message.FromId);
-			}
-			else
-			{
-				unsuspendedFor.Remove(message.FromId);
-			}
-
-			var newOwner = unsuspendedFor.Count != 0 ? unsuspendedFor[0] : uint.MaxValue;
-			SetAuthority(message.Identity, newOwner);
-		}
-
-		private static void SetAuthority(QNetworkIdentity identity, uint id)
-		{
-			var oldConn = identity.ClientAuthorityOwner;
-			var newConn = id != uint.MaxValue
-				? QNetworkServer.connections.First(x => x.GetPlayerId() == id)
-				: null;
-
-			if (oldConn == newConn)
-			{
-				return;
-			}
-
-			if (oldConn != null)
-			{
-				identity.RemoveClientAuthority(oldConn);
-			}
-
-			if (newConn != null)
-			{
-				identity.AssignClientAuthority(newConn);
-			}
-
-			DebugLog.DebugWrite($"{QSBPlayerManager.LocalPlayerId}.{identity.NetId}:{identity.gameObject.name} - set authority to {id}");
+			SuspendableManager.SetSuspended(message.FromId, message.Identity, message.Suspended);
 		}
 	}
 }
